feat

+++ database/migrations/2026_01_16_132044_remove_observations_from_cost_center_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('application_order_cost_center', function (Blueprint $table) {
-            $table->dropColumn('observations');
+            if (Schema::hasColumn('application_order_cost_center', 'observations')) {
+                $table->dropColumn('observations');
+            }
         });
     }
